% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{feature.prevalence}
\alias{feature.prevalence}
\title{Feature prevalence}
\usage{
feature.prevalence(li, index, hist.plot = TRUE)
}
\arguments{
\item{li}{the list result of \code{feature.removal}.}

\item{index}{removing the features of the first \code{index} iterations. It
allows a positive integer or a proper fraction. If inproper fraction, it is
regarded as \code{as.integer(index)}.}

\item{hist.plot}{bool. A switch to plot the histogram of the remaining features.}
}
\value{
Table
}
\description{
Compute the feature prevalence after removing the features of
the first \code{index} iterations.
}
\examples{

g1 = SWR1[,5:ncol(SWR1)]
g0 = SWR0[,5:ncol(SWR0)]

row.names(g1) <- SWR1[,4]
row.names(g0) <- SWR0[,4]

result.complex <- feature.removal(g1, g0,
    cutoff1=0.95, cutoff0=0.925,
    offset=c(0.5, 1, 2))

# index is a proportion in 0-1
prevalence.result <- feature.prevalence(result.complex, 0.5, hist.plot=TRUE)

# index is a positive integer
prevalence.result <- feature.prevalence(result.complex, 233, hist.plot=TRUE)
}
