% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{feature_hist}
\alias{feature_hist}
\title{Plot histogram of feature prevalence}
\usage{
feature_hist(li, index)
}
\arguments{
\item{li}{the list result of \code{feature_removal}.}

\item{index}{removing the features of the first \code{index} iterations. It
allows a positive integer or a proper fraction. If inproper fraction, it is
regarded as \code{as.integer(index)}.}
}
\value{
histogram
}
\description{
Compute the feature prevalence (present in different cutoffs)
after removing the features of
the first \code{index} iterations, and then plot the histogram of remaining
features. It calls \code{feature_prevalence(..., hist.plot=TRUE)}.
}
\examples{
g1 <- SWRG1; g0 <- SWRG0

result.complex <- feature_removal(g1, g0,
    cutoff1=0.95, cutoff0=0.925,
    offset=c(0.5, 1, 2))

# index is a proportion in 0-1
feature_hist(result.complex, 0.5)

# index is a positive integer
feature_hist(result.complex, 233)
}
