% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{feature_prevalence}
\alias{feature_prevalence}
\title{Feature prevalence}
\usage{
feature_prevalence(li, index, hist.plot = TRUE)
}
\arguments{
\item{li}{the list result of \code{feature_removal}.}

\item{index}{removing the features of the first \code{index} iterations. It
allows a positive integer or a proper fraction. If inproper fraction, it is
regarded as \code{as.integer(index)}.}

\item{hist.plot}{bool. A switch to plot the histogram of the remaining
features.}
}
\value{
Matrix
}
\description{
Compute the feature prevalence after removing the features of
the first \code{index} iterations.
}
\examples{
g1 <- SWRG1; g0 <- SWRG0

result.complex <- feature_removal(g1, g0,
    cutoff1=0.95, cutoff0=0.925,
    offset=c(0.5, 1, 2))

# index is a proportion in 0-1
prevalence.result <- feature_prevalence(result.complex, 0.5, hist.plot=TRUE)

# index is a positive integer
prevalence.result <- feature_prevalence(result.complex, 233, hist.plot=TRUE)
}
