% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{ggiteration_trace}
\alias{ggiteration_trace}
\title{Iteration trace of removed scores}
\usage{
ggiteration_trace(li)
}
\arguments{
\item{li}{the list result of \code{feature.removal}.}
}
\value{
ggplot2 object.
}
\description{
plot the score of removed feature in each iteration.
}
\examples{
g1 <- SWRG1; g0 <- SWRG0

result.complex <- feature.removal(g1, g0,
    cutoff1=0.95, cutoff0=0.925,
    offset=c(0.5, 1, 2))

# it is a ggplot2 object, so plus sign is available
ggiteration_trace(result.complex) + theme_bw()
}
