% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{linear.combine}
\alias{linear.combine}
\title{Score combine methods linear combine}
\usage{
linear.combine(g1.score.feature, g0.score.feature, offset = 1)
}
\arguments{
\item{g1.score.feature}{feature score vector for g1.}

\item{g0.score.feature}{feature score vector for g0.}

\item{offset}{adjusts the proportion of \code{g1.score.feature}. Default is \code{1}.}
}
\value{
\code{g1.score.feature * offset + g0.score.feature}.
}
\description{
To combine the feature score vectors of g1 and g0. The method
used in \code{feature.removal} must have three parameters in order,
\code{g1.score.feature}, \code{g0.score.feature}, and \code{offset}.
See details in the help page of function \code{feature.removal}.
}
\examples{
linear.combine(0.2, 0.3, 2)
linear.combine(1:2, 3:4, 1)
}
