% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{min_max}
\alias{min_max}
\title{Score standardization methods min max}
\usage{
min_max(x, na.rm = TRUE)
}
\arguments{
\item{x}{The first parameter of score standardization method used in
\code{feature.removal} is the sum-up dataframe. See details in the help page
of function \code{feature.removal}.}

\item{na.rm}{Bool. Remove NA or not.}
}
\value{
Numeric table
}
\description{
Normalizing the vector \code{x} to 0-1 range.
See details in the help page of function \code{feature.removal}.
}
\examples{
min_max(1:5)
min_max(c(1:5, NA), na.rm=TRUE)
}
