% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_removal.R
\name{multiple.feature.removal.iteration}
\alias{multiple.feature.removal.iteration}
\title{Stepwise feature removal method internal wrapper}
\usage{
multiple.feature.removal.iteration(g1.signal, g0.signal, offset = 1,
  weight.method = "reciprocal.colSums",
  scoreStandardization.method = "min_max",
  scoreCombine.method = "linear.combine", ...)
}
\arguments{
\item{g1.signal}{a dataframe with the row of feature, and the column of observation.
Cells are numeric or bool.}

\item{g0.signal}{a dataframe with the same row names as \code{g1.signal}. Normally, the observations
in \code{g0.signal} are in the distinct group of \code{g1.signal}.}

\item{offset}{a parameter in \code{scoreCombine.method}. It adjusts the score proportion
of g1 and g2. Besides, \code{offset} can be a number or a numeric vector.
If it is a vector, the overall iteration is done for each offset respectively.
See more in parameter \code{scoreCombine.method}.}

\item{weight.method}{\code{gx.weight}, weight of gx, is computed using
\code{weight.method}. The weight is for the observations/columns, not the
features/rows. The default weight method is \code{reciprocal.colSums}, ie.
\code{1 / (1 + colSums(gx.signal, na.rm=T))}. You can specify your own function,
and the first parameter of the function should be the exact word of \code{gx.signal}.}

\item{scoreStandardization.method}{Default standardization method is Min-Max,
ie. normalizing the vector to 0-1 range. You can specify your own function,
and the first parameter of the function is the sum-up dataframe. See more in
Details section.}

\item{scoreCombine.method}{to combine the feature score vectors of g1 and g0.
This method must have three parameters in order, \code{g1.score.feature},
\code{g0.score.feature}, and \code{offset}. Default method is \code{linear.combine}.
\code{offset} in the default method adjusts the proportion of \code{g1.score.feature}.
Specifically, \code{g1.score.feature * offset + g0.score.feature}. Besides,
\code{offset} can be a number or a vector. If it is a vector, the overall
iteration is done for each offset respectively.}

\item{...}{Other parameter passed to method of expression class.}
}
\value{
a list with names "offset", "removed.feature_names", "removed.scores", and "max.scores".
}
\description{
This function screens features iteratively in consideration of
limiting overfitting and overall performance.
}
\details{
See details in \code{feature.removal}
}
\seealso{
Other iteration: \code{\link{feature.removal.iteration}},
  \code{\link{feature.removal}}
}
