% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{ones}
\alias{ones}
\title{Weight methods ones}
\usage{
ones(gx.signal)
}
\arguments{
\item{gx.signal}{The first parameter of the weight method used in
\code{feature.removal} must be the exact word \code{gx.signal}. See details
in the help page of function \code{feature.removal}.}
}
\value{
\code{1}.
}
\description{
See details in the help page of function \code{feature.removal}.
}
\examples{
1 == ones(2)
1 == ones(c(4,6))
}
\seealso{
Other weightmethod: \code{\link{reciprocal.colSums}}
}
