% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{reciprocal.colSums}
\alias{reciprocal.colSums}
\title{Weight methods reciprocal colSums}
\usage{
reciprocal.colSums(gx.signal)
}
\arguments{
\item{gx.signal}{The first parameter of the weight method used in
\code{feature.removal} must be the exact word \code{gx.signal}.}
}
\value{
\code{1 / (1 + colSums(gx.signal, na.rm=T))}.
}
\description{
See details in the help page of function \code{feature.removal}.
}
\examples{
gx.singal <- data.frame(x=1:5, t=2:6, k=c(3:6, NA))
reciprocal.colSums(gx.singal)
}
\seealso{
Other weightmethod: \code{\link{ones}}
}
