% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imports.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{\%>\%}
\alias{reexports}
\alias{hist}
\alias{reexports}
\alias{stack}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{graphics}{\code{\link[graphics]{hist}}}

  \item{magrittr}{\code{\link[magrittr]{\%>\%}}}

  \item{utils}{\code{\link[utils]{stack}}}
}}

