% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{score_standardization_methods}
\alias{score_standardization_methods}
\alias{min_max}
\title{Score standardization methods}
\usage{
min_max(x, na.rm = TRUE)
}
\arguments{
\item{x}{The first parameter of score standardization method used in
\code{feature.removal} is the sum-up dataframe. See details in the help page
of function \code{feature.removal}.}

\item{na.rm}{Bool. Remove NA or not.}
}
\value{
Numeric, normalized \code{x} to 0-1 range.
}
\description{
Score standardization methods of function 
\code{feature.removal}.
}
\examples{
min_max(1:5)
min_max(c(1:5, NA), na.rm=TRUE)
}
