% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{SummarizedData}
\alias{SummarizedData}
\title{Summarized data}
\format{RangedSummarizedExperiment.}
\usage{
SummarizedData
}
\description{
RangedSummarizedExperiment object contains the probability data of the control and positive samples. It is the integration of \code{SWRG0} and \code{SWRG1}.
}
\seealso{
Other data: \code{\link{SWRG0}}, \code{\link{SWRG1}}
}
\keyword{datasets}
