% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{SWR1}
\alias{SWR1}
\title{Group 1 data}
\format{A data frame with 26 variables: \code{Chr}, \code{Start},
  \code{End}, \code{Name}, \code{G0Sample1}, \code{G0Sample2}, ...,
  \code{G0Sample22}.}
\usage{
SWR1
}
\description{
Probability data of positive samples.
}
\keyword{datasets}
