% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{SWRG0}
\alias{SWRG0}
\title{Group 0 data}
\format{A data frame with 22 variables: \code{G0Sample1}, \code{G0Sample2}, ...,
  \code{G0Sample22}.}
\usage{
SWRG0
}
\description{
Probability data of control samples.
}
\seealso{
Other data: \code{\link{SWRG1}},
  \code{\link{SummarizedData}}
}
\keyword{datasets}
