% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{SWRG1}
\alias{SWRG1}
\title{Group 1 data}
\format{A data frame with 22 variables: \code{G0Sample1}, \code{G0Sample2}, ...,
  \code{G0Sample22}.}
\usage{
SWRG1
}
\description{
Probability data of positive samples.
}
\seealso{
Other data: \code{\link{SWRG0}},
  \code{\link{SummarizedData}}
}
\keyword{datasets}
