% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{weight_methods}
\alias{weight_methods}
\alias{reciprocal_colSums}
\alias{ones}
\title{Weight methods}
\usage{
reciprocal_colSums(gx.signal)

ones(gx.signal)
}
\arguments{
\item{gx.signal}{The first parameter of the weight method used in
\code{feature.removal} must be the exact word \code{gx.signal}.}
}
\value{
\code{1 / (1 + colSums(gx.signal, na.rm=T))}.

\code{1}.
}
\description{
Weight methods of function \code{feature.removal}.
}
\examples{
gx.singal <- data.frame(x=1:5, t=2:6, k=c(3:6, NA))
reciprocal_colSums(gx.singal)
1 == ones(2)
1 == ones(c(4,6))
}
