% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{feature.screen}
\alias{feature.screen}
\title{Screening feature using prevalence}
\usage{
feature.screen(features, prevalence)
}
\arguments{
\item{features}{result of \code{feature.prevalence(...)}}

\item{prevalence}{the prevalence cutoff of features. The features with
prevalence less than \code{prevalence} are removed.}
}
\value{
Vector
}
\description{
Return the screened feature names.
}
\examples{
data(SWRG1); data(SWRG0)
g1 <- SWRG1; g0 <- SWRG0

result.complex <- feature.removal(g1, g0,
    cutoff1=0.95, cutoff0=0.925,
    offset=c(0.5, 1, 2))

prevalence.result <- feature.prevalence(result.complex, 233, hist.plot=TRUE)

feature.list <- feature.screen(prevalence.result, 3)
}
