% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{SummarizedData}
\alias{SummarizedData}
\title{Summarized data}
\format{RangedSummarizedExperiment.}
\usage{
SummarizedData
}
\description{
RangedSummarizedExperiment object contains the probability
data of the control and positive samples. It is the integration of
\code{SWRG0} and \code{SWRG1}. \code{colData} in \code{SummarizedData}
contains a column 'Group' indicating normal(0) from malignant(1). In assay,
column represents samples,
row represents genome ranges related to a disease.
}
\details{
We identified 299 genomic regions related to the methylation status
of a disease. Then, we sequenced the cfDNA of 44 subjects. 22 were malignant,
 while others were normal. We built a statistical model to compute the
 probability of the disease for each region and subject, and this dataframe
 contains the probabilities of 299 regions and 44 samples.
}
\seealso{
Other data: \code{\link{SWRG0}}, \code{\link{SWRG1}}
}
\keyword{datasets}
