% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{SWRG1}
\alias{SWRG1}
\title{Group 1 data}
\format{A data frame with 22 variables: \code{G0Sample1},
\code{G0Sample2}, ..., \code{G0Sample22}.}
\usage{
SWRG1
}
\description{
Probability data of positive samples. Column represents samples,
row represents genome ranges related to a disease.
}
\details{
We identified 299 genomic regions related to the methylation status
of a disease. Then, we sequenced the cfDNA of 44 subjects. 22 were malignant,
 while others were normal. We built a statistical model to compute the
 probability of the disease for each region and subject, and this dataframe
 contains the probabilities of 299 regions and 22 malignant samples.
}
\seealso{
Other data: \code{\link{SWRG0}},
  \code{\link{SummarizedData}}
}
\keyword{datasets}
